requirements.txt
# Core shared infrastructure dependencies
python-dotenv>=1.0.0
duckdb>=0.9.0

# MCP and photo-migration dependencies
mcp>=0.1.0
playwright>=1.40.0
Pillow>=10.0.0

# Google API dependencies (for Phase 2+)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Utility dependencies
aiofiles>=23.0.0
tenacity>=8.2.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0